feat: update device registration methods and refactor probe handling
