+++ rdrive/src/driver/mod.rs
@@ -1,5 +1,7 @@
 use rdif_base::DriverGeneric;
 
+use crate::Descriptor;
+
 #[macro_use]
 mod _macros;
 
@@ -13,6 +15,27 @@ impl DriverGeneric for Empty {
     fn close(&mut self) -> Result<(), rdif_base::KError> {
         Ok(())
     }
+}
+
+pub struct PlatformDevice {
+    pub descriptor: Descriptor,
+}
+
+impl PlatformDevice {
+    pub(crate) fn new(descriptor: Descriptor) -> Self {
+        Self { descriptor }
+    }
+
+    /// Register a device to the driver manager.
+    ///
+    /// # Panics
+    /// This method will panic if the device with the same ID is already added
+    pub fn register<T: DriverGeneric>(self, driver: T) {
+        crate::edit(|manager| {
+            manager.dev_container.insert(self.descriptor, driver);
+        });
+    }
+
 }
 
 def_driver_rdif!(Intc);
